Registration completed

+++ app/Http/Controllers/ProductDetailController.php
@@ -16,6 +16,7 @@ use App\Category;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Redirect;
 
 class ProductDetailController extends Controller
 {
@@ -23,7 +24,7 @@ class ProductDetailController extends Controller
     {
         $cake = Cake::with('cakedesc')->find($id);
         $cake_type = Cake::where('type', '=', $category)->get();
-        $category = Category::all();
+        $category =$category = Category::where('seperator','=',1)->get();;
         return view('productdetails', ['cake' => $cake, 'categories' => $category, 'cakewithtype' => $cake_type]);
     }
 
@@ -31,53 +32,11 @@ class ProductDetailController extends Controller
     public function addtoCart()
     {
         // This will add items to cart
-      if (Auth::check()) {
-
-               $cake_desc_id = request()->input('CakeDescId');
-               $qty = request()->input('Quantity');
-               $wording = request()->input('Wording');
-
-
-            //status will be 1 if the user has add to the cart and not purchased
-           $cart = Cart::where('user_id', '=', Auth::User()->id)->where("status", '=', 1)->first();
-            $cake_desc = CakeDesc::find($cake_desc_id);
-            if ($cart == null) {
-                $cart = new Cart();
-                $cart->req_date = Carbon::today()->toDateString();
-                $cart->sub_total = 0;
-                $cart->ord_date = Carbon::today()->toDateString();
-                $cart->sms_state = 1;
-                $cart->user_id = Auth::User()->id;
-                $cart->status = 1;
-                $cart->save();
-
-                $cartitem = new CartItem();
-                $cartitem->qty = $qty;
-                $cartitem->wording = $wording;
-                $cartitem->price = ((int)$qty) * ($cake_desc->price);
-                $cartitem->cake_desc_id = $cake_desc_id;
-                $cart->cartitem()->save($cartitem);
-
-            } else {
-                $cartitem = new CartItem();
-                $cartitem->qty = $qty;
-                $cartitem->wording = $wording;
-                $cartitem->price = ((int)$qty) * ($cake_desc->price);
-                $cartitem->cake_desc_id = $cake_desc_id;
-                $cart->cartitem()->save($cartitem);
-            }
-
-      //      return view('/');
-            return response()->json(['retvalue' => $cake_desc_id]);
-       } else {
-          /* $cake_desc_id = request()->input('CakeDescId');
-           $qty = request()->input('Quantity');
-           $wording = request()->input('Wording');
-           $req->session()->put('cake_desc_id',$cake_desc_id);
-           $req->session()->put('qty',$qty);
-           $req->session()->put('wording',$wording);*/
-            return response()->json(['retvalue' => 'login']);
-       }
+        $cake_desc_id = request()->input('CakeDescId');
+        $qty = request()->input('Quantity');
+        $wording = request()->input('Wording');
+        $this->commonCart($cake_desc_id, $qty, $wording);
+        return response()->json(['retvalue' => $cake_desc_id]);
 
     }
 
@@ -106,9 +65,48 @@ class ProductDetailController extends Controller
 
     public function removeItem($order_id)
     {
-        $cartitem=CartItem::find($order_id);
-        $cartitem->delete();
+        $cartitem = CartItem::find($order_id);
+        if ($cartitem != null) {
+            $cartitem->delete();
+        }
         return $this->viewCart();
     }
 
+    public function addtoCartWithId($id, $qty, $wording = null)
+    {
+        $this->commonCart($id, $qty, $wording);
+        return Redirect::to('/cart');
+    }
+
+    public function commonCart($cake_desc_id, $qty, $wording)
+    {
+        //status will be 1 if the user has add to the cart and not purchased
+        $cart = Cart::where('user_id', '=', Auth::User()->id)->where("status", '=', 1)->first();
+        $cake_desc = CakeDesc::find($cake_desc_id);
+        if ($cart == null) {
+            $cart = new Cart();
+            $cart->req_date = Carbon::today()->toDateString();
+            $cart->sub_total = 0;
+            $cart->ord_date = Carbon::today()->toDateString();
+            $cart->sms_state = 1;
+            $cart->user_id = Auth::User()->id;
+            $cart->status = 1;
+            $cart->save();
+
+            $cartitem = new CartItem();
+            $cartitem->qty = $qty;
+            $cartitem->wording = $wording;
+            $cartitem->price = ((int)$qty) * ($cake_desc->price);
+            $cartitem->cake_desc_id = $cake_desc_id;
+            $cart->cartitem()->save($cartitem);
+
+        } else {
+            $cartitem = new CartItem();
+            $cartitem->qty = $qty;
+            $cartitem->wording = $wording;
+            $cartitem->price = ((int)$qty) * ($cake_desc->price);
+            $cartitem->cake_desc_id = $cake_desc_id;
+            $cart->cartitem()->save($cartitem);
+        }
+    }
 }
